fix: breadcrumb  :id path
