+++ src/hooks/useAuth.ts
@@ -1,5 +1,5 @@
 import { loginService } from "@/services/auth.service";
-import { snackBarElement } from "@/helpers/snackBarElement";
+import { snackBarElement } from "@/utils/snackBarElement";
 import { UserLogin } from "@/interfaces";
 import { useAuthStore } from "@/store/authStore";
