+++ src/pages/Login.page.tsx
@@ -1,6 +1,6 @@
 import { Box, Typography } from "@mui/material";
 import { FormLogin } from "@/components/FormLogin";
-import { AuthLayout } from "../layouts/AuthLayout";
+import { AuthLayout } from "@/layouts/AuthLayout";
 
 export const LoginPage = () => {
   return (
