+++ src/utils/routeName.ts
@@ -1,0 +1,10 @@
+export const routeNames: { [key: string]: string } = {
+  "/usuarios": "Usuarios",
+  "/usuarios/nuevo": "Usuarios > Nuevo",
+  "/roles": "Roles",
+  "/permisos": "Permisos",
+  "/productos": "Productos",
+  "/productos/nuevo": "Productos > Nuevo",
+  "/productos/editar/:id": "Productos > Editar",
+  "/categorias": "Categorias",
+};
